+++ internal/cli/status.go
@@ -30,6 +30,27 @@ func newStatusCommand(global *globalOptions) *cobra.Command {
 				return printJSON(release)
 			}
 			fmt.Printf("Name: %s\nStatus: %s\nRevision: %d\nPackage: %s@%s\nApp: %s\nCompose Project: %s\n", release.Name, release.Status, release.Revision, release.PackageName, release.PackageVersion, release.AppVersion, release.ComposeProject)
+			if release.Parent != nil {
+				if release.Parent.Alias != "" {
+					fmt.Printf("Parent: %s (dependency %s as %s)\n", release.Parent.Name, release.Parent.DependencyName, release.Parent.Alias)
+				} else {
+					fmt.Printf("Parent: %s (dependency %s)\n", release.Parent.Name, release.Parent.DependencyName)
+				}
+			}
+			if len(release.Dependencies) > 0 {
+				fmt.Println("Dependencies:")
+				for _, dep := range release.Dependencies {
+					alias := ""
+					if dep.Alias != "" {
+						alias = " as " + dep.Alias
+					}
+					pkg := dep.PackageName
+					if dep.PackageVersion != "" {
+						pkg = pkg + "@" + dep.PackageVersion
+					}
+					fmt.Printf("  - %s%s -> %s (%s, %s)\n", dep.Name, alias, dep.Release, pkg, dep.Status)
+				}
+			}
 			if release.DockyardVersion != "" {
 				fmt.Printf("Dockyard Version: %s\n", release.DockyardVersion)
 			}